+++ android/src/main/kotlin/it/mainella/audible_mode/utils/audible_profile.kt
@@ -1,5 +1,6 @@
-internal enum class AudibleProfile(val rawValue: String) {
+enum class AudibleProfile(val rawValue: String) {
     SILENT_MODE("SILENT_MODE"),
     VIBRATE_MODE("VIBRATE_MODE"),
     NORMAL_MODE("NORMAL_MODE"),
+    UNDEFINED("UNDEFINED"),
 }