+++ android/src/main/kotlin/it/mainella/audible_mode/utils/constants.kt
@@ -2,7 +2,11 @@ package it.mainella.audible_mode.utils
 
 class Constants {
     companion object {
-        val METHOD_CHANNEL: String = "CHANNEL_AUDIBLE_MODE";
-        val GET_AUDIBLE_MODE: String = "GET_AUDIBLE_MODE";
+        const val METHOD_CHANNEL: String = "CHANNEL_AUDIBLE_MODE"
+        const val EVENT_CHANNEL: String = "AUDIBLE_MODE_STREAM"
+        const val GET_AUDIBLE_MODE: String = "GET_AUDIBLE_MODE"
+        const val GET_CURRENT_VOLUME: String = "GET_CURRENT_VOLUME"
+        const val SET_VOLUME: String = "SET_VOLUME"
+        const val GET_MAX_VOLUME: String = "GET_MAX_VOLUME"
     }
 }