+++ ios/Classes/utils/audible_profile.swift
@@ -5,7 +5,7 @@
 //  Created by Andrea Mainella on 30/07/21.
 //
 
-enum AudibleMode : String {
-  case SILENT_MODE
-  case NORMAL_MODE
+enum AudibleProfile : String {
+    case SILENT_MODE
+    case NORMAL_MODE
 }