+++ ios/Classes/utils/constants.swift
@@ -7,5 +7,9 @@
 
 class Constants{
     static let METHOD_CHANNEL = "CHANNEL_AUDIBLE_MODE";
+    static let EVENT_CHANNEL = "AUDIBLE_MODE_STREAM";
     static let GET_AUDIBLE_MODE = "GET_AUDIBLE_MODE";
+    static let GET_CURRENT_VOLUME = "GET_CURRENT_VOLUME";
+    static let SET_VOLUME = "SET_VOLUME";
+    static let GET_MAX_VOLUME = "GET_MAX_VOLUME";
 }