+++ lib/audible_mode.dart
@@ -1,9 +1,4 @@
 library audible_mode;
 
-import 'package:enum_to_string/enum_to_string.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-part 'src/audible.dart';
-part 'utils/audible_profile.dart';
-part 'utils/constants.dart';
+export 'src/audible.dart' show Audible;
+export 'src/utils/audible_profile.dart' show AudibleProfile;