+++ lib/src/utils/audible_profile.dart
@@ -1,0 +1,8 @@
+/// These are the system audio profiles
+/// in iOS there are only: [AudibleProfile.SILENT_MODE] and [AudibleProfile.NORMAL_MODE]
+enum AudibleProfile {
+  SILENT_MODE,
+  VIBRATE_MODE,
+  NORMAL_MODE,
+  UNDEFINED,
+}