+++ lib/src/utils/constants.dart
@@ -1,0 +1,9 @@
+/// This class contain Method and Event channel name
+class Constants {
+  static const String METHOD_CHANNEL = 'CHANNEL_AUDIBLE_MODE';
+  static const String EVENT_CHANNEL = 'AUDIBLE_MODE_STREAM';
+  static const String GET_AUDIBLE_PROFILE = 'GET_AUDIBLE_MODE';
+  static const String GET_CURRENT_VOLUME = 'GET_CURRENT_VOLUME';
+  static const String SET_VOLUME = 'SET_VOLUME';
+  static const String GET_MAX_VOLUME = 'GET_MAX_VOLUME';
+}