+++ lib/utils/audible_profile.dart
@@ -1,8 +1,0 @@
-part of audible_mode;
-
-enum AudibleProfile {
-  SILENT_MODE,
-  VIBRATE_MODE,
-  NORMAL_MODE,
-  UNDEFINED,
-}